pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.2",
    "flask>=3.1.2",
    "imageio-ffmpeg>=0.6.0",
    "requests>=2.32.5",
    "yt-dlp>=2025.8.22",
]
